+++ lib/project_types/extension/models/product.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module Extension
+  module Models
+    class Product
+      include SmartProperties
+
+      property :variant_id, accepts: Integer, converts: :to_i
+      property :quantity, accepts: Integer, default: 1
+    end
+  end
+end