+++ test/project_types/extension/features/runtimes/admin_test.rb
@@ -15,6 +15,7 @@ module Extension
             :port,
             :public_url,
             :renderer_version,
+            :resource_url,
             :shop,
             :uuid,
           ]